Member activity reports (#93)
This is an early draft PR of a first revision of a feature to support
the Reactiflux Star Helper program — tho it will be useful for other
types of roles as well! That's the whole point of automating this.

<img width="567" alt="Screenshot 2024-12-16 at 2 15 19 AM"
src="https://github.com/user-attachments/assets/de4e3cad-daba-44eb-9539-78e0bebe1b22"
/>

This is ~ a proof of concept to get a query together for examining the
data, which it accomplishes. There's a lot lot lot of supporting code
needed to turn this into "a feature", my intentions with this PR are to
get enough code together to support the 2024 Q4 Star Helper program,
then productize in the next few months.

Some examples of work that will need to be done before this is a
production feature:

- Add configurable channels/categories, and intervals
- Flexible scoring criteria
- Clearer details around data collection (privacy policy — I believe
we're very light here, as we don't store any content, just metadata)
- Inline scoring/member review? (I don't think it's feasible to pull in
user history here, as I don't want to store it and Discord doesn't make
it convenient to retrieve)
- Exploration into the best reporting options (inline discord messages
from the bot? attached PDFs/PNGs? web app?)