Upgrade to react-router@7, eslint@9, version bump the rest (#101)
Goal here is to get fully up-to-date with modern Remix/React Router.
I've gotten this working ~ correctly (as best I can tell right now) with
hot reloading for server/Remix/client (which are different things! the
Express server and Remix SSR are built separately, it's a PITA).

This is a draft PR while I continue working on moving to React Router v7
(from Remix v2).