Tighten up Track output (#183)
This is the new format
<img width="550" height="188" alt="Screenshot 2025-09-27 at 10 28 24 PM"
src="https://github.com/user-attachments/assets/f4d6613e-5922-46cb-b966-893a97f94308"
/>

I like it a lot better. It doesn't totally flood the screen when spam
hits, which should be a great boon. It still quotes the entirety of the
post in the log thread.