Adjust environment variables for production payments, improve error logging (#217)
Payments work locally, but fail in production for some reason. All the
values appear to be in place correctly, but the logging currently only
shows `{}` for the error, which isn't super duper helpful here. This PR:

- fixes some poor env practices
- adds a missing webhook secret (which shouldn't have been able to cause
the observed failure, but hey it's probably wrong too)
- improves error logging across the app, which should at least provide
more detail about what's going wrong if this PR fails to remedy the
issue