Pull in Star Helper cohort analysis work from DFG (#226)
I meant to build more on top of this before merging, but this is now
beginning to risk rework due to merge conflicts and it's in a workable
state, so I'm going to merge. Thank you @DanielFGray for the bulk of
this work!

The Star Hunter feature is going to be gated at launch as a future
premium feature — it currently relies on hardcoded channels from
Reactiflux, which obviously is not workable as a production feature, and
revising it to handle advance configuration/work delegation is more
effort than I want to gate further development on.

So tl;dr
- Star Hunter feature works for Reactiflux (only)
- Intended in the future as a premium add-on feature, available to paid
plans
- Merging this so it can continue to be built upon

---------

Co-authored-by: DanielFGray <danielfgray@gmail.com>